requirements.txt
PyMySQL==1.1.0
DBUtils==3.0.2
python-dotenv==1.0.0
cryptography==41.0.5
pytest==7.4.0 